--- src/main/java/util/validator/NumberValidator.java
@@ -1,13 +1,0 @@
-package util.validator;
-
-public class NumberValidator {
-
-    private static void validateNumberFormat(String input) {
-        try {
-            Integer.parseInt(input);
-        } catch (NumberFormatException e) {
-            throw new NumberFormatException("숫자가 아닌 문자열이 입력되었습니다.");
-        }
-    }
-
-}
